+++ cypress/locators/accordion/index.js
@@ -4,16 +4,7 @@ import {
   ACCORDION_ICON,
 } from "./locators";
 
-// component preview locators
-export const accordionIframe = () => cy.iFrame(ACCORDION_PREVIEW);
-export const accordionTitleContainerInIframe = () =>
-  accordionIframe().find(ACCORDION_TITLE_CONTAINER);
-export const accordionIcon = () =>
-  accordionTitleContainerInIframe().find(ACCORDION_ICON);
-export const accordionTitleContainerByPositionInIfame = (index) =>
-  accordionIframe().find(ACCORDION_TITLE_CONTAINER).eq(index).children();
-
-// NoIFrame locators
+// locators
 export const accordionTitleContainer = () =>
   cy.get(ACCORDION_PREVIEW).find(ACCORDION_TITLE_CONTAINER);
 export const accordionDefaultTitle = () => cy.get(ACCORDION_TITLE_CONTAINER);
@@ -23,3 +14,5 @@ export const accordionTitleContainerByPosition = (index) =>
     .find(ACCORDION_TITLE_CONTAINER)
     .eq(index)
     .children();
+export const accordionIcon = () =>
+  accordionTitleContainer().find(ACCORDION_ICON);