+++ cypress/support/step-definitions/badge-steps.js
@@ -1,11 +1,7 @@
-import {
-  badge,
-  badgeCounterNoIFrame,
-  badgeNoIFrame,
-} from "../../locators/badge";
+import { badge, badgeCounter } from "../../locators/badge";
 
 Then("Badge component counter is set to {int}", (value) => {
-  badgeCounterNoIFrame()
+  badgeCounter()
     .invoke("show")
     .should("be.visible")
     .invoke("text")
@@ -13,11 +9,11 @@ Then("Badge component counter is set to {int}", (value) => {
 });
 
 Then("Badge component counter does not exist", () => {
-  badgeNoIFrame().should("not.exist");
+  badge().should("not.exist");
 });
 
 When("I focus onto Badge component", () => {
-  badgeNoIFrame().focus();
+  badge().focus();
 });
 
 When("I click onto Badge component", () => {
@@ -25,7 +21,7 @@ When("I click onto Badge component", () => {
 });
 
 Then("Badge component cross icon has proper color", () => {
-  badgeNoIFrame()
+  badge()
     .should("have.css", "background")
     .then(($el) => {
       expect($el).contains("rgb(0, 129, 93)");